Convert hour to an integer

+++ library/src/main/java/com/alamkanak/weekview/drawing/HeaderRowDrawer.java
@@ -464,7 +464,7 @@ public class HeaderRowDrawer {
 
         void goToDate(Calendar date);
 
-        void goToHour(double hour);
+        void goToHour(int hour);
 
     }
